requirements-unittest.txt
openmeteo_requests>=1.3.0,<2
#pandas>=2.2.3,<3
paho-mqtt>=2.1.0
pytz>=2024.2
requests>=2.32.3
requests_cache>=1.2.1
retry_requests>=2.0.0
unittest
